Moved parts of getaddrinfo into safe wrapper function. Better handling of gai errors in agent->layer.
